+++ src/instance_proc_addr_loader.rs
@@ -153,6 +153,7 @@ gen_instance_proc_addr_loader!(
     "nv_external_memory_win32_1" => nv_external_memory_win32: NV_external_memory_win32 [fn load_nv_external_memory_win32],
     "nvx_device_generated_commands_1" => nvx_device_generated_commands: NVX_device_generated_commands [fn load_nvx_device_generated_commands],
     "khr_get_physical_device_properties2_1" => khr_get_physical_device_properties2: KHR_get_physical_device_properties2 [fn load_khr_get_physical_device_properties2],
+    "khr_maintenance1_1" => khr_maintenance1: KHR_maintenance1 [fn load_khr_maintenance1],
 );
 
 addr_proc_struct!(CoreNullInstance {
@@ -411,4 +412,9 @@ addr_proc_struct!(KHR_get_physical_device_properties2 {
     pfn vkGetPhysicalDeviceQueueFamilyProperties2KHR: PFN_vkGetPhysicalDeviceQueueFamilyProperties2KHR,
     pfn vkGetPhysicalDeviceMemoryProperties2KHR: PFN_vkGetPhysicalDeviceMemoryProperties2KHR,
     pfn vkGetPhysicalDeviceSparseImageFormatProperties2KHR: PFN_vkGetPhysicalDeviceSparseImageFormatProperties2KHR,
+});
+
+#[cfg(feature = "khr_maintenance1_1")]
+addr_proc_struct!(KHR_maintenance1 {
+    pfn vkTrimCommandPoolKHR: PFN_vkTrimCommandPoolKHR,
 });