+++ src/nvx_device_generated_commands.rs
@@ -195,6 +195,7 @@ pub struct VkObjectTableIndexBufferEntryNVX {
     pub type_: VkObjectEntryTypeNVX,
     pub flags: VkObjectEntryUsageFlagsNVX,
     pub buffer: VkBuffer,
+    pub indexType: VkIndexType,
 }
 
 #[repr(C)]