+++ CHANGELOG.md
@@ -6,6 +6,7 @@
  * Support revision 4 of `VK_EXT_debug_report`
  * Add extension `VK_KHR_get_physical_device_properties2` (revision 1)
  * Add extension `VK_KHR_shader_draw_parameters` (revision 1)
+ * Add extension `VK_KHR_maintenance1` (revision 1)
 
 ## Version 0.8.1, released on 29.03.2017